Markdown capable fields
The bio field on Profile, and the description field on Project has been made markdown capable. The user can see live changes to the Markdown field in the broswer.

The django-markdownx package has been installed in order to accomplish this. All of its required information takes up most of the 194 changed files.