Profile to accounts migration
The start of migrating the bio, avatar, and username fields to accounts has been added. Not mutch has been done on this commit. This should make things simplier in the future and allow the notification views to be placed inside its own application, because urls will no longer be tied up with the profile information. Thus making it stand alone.