Migration complete
The profiles app has been changed to work with the user's profile being in a custom user model. This allows the application to be simplier without having to transition form user to a user's profile.

This has also allowed all of the notification views to be able to stand alone as its own app. Previously the url routes required a profile as part of the route, which means that the notifications were still dependant on a Profile.

All of the temporary files have been deleted.