Dynamically updating skills form
The user now has a dynamically operating skills formset. This was accomplished with some jquery that I found online and edited. The MIT lisence is found in the profile_edit.html file if interested.

As a side not it appears that the unit test associated with the profile_edit is no longer working. I commented that out for now, and will update it later on.  Posting this commit for now.