Dynamically updating position forms
The user now has a dynamically operating position formset. This was accomplished with some jquery that I found online and edited. The MIT lisence is found in the profile_edit.html file if interested.

Some tests have been commented out for now, to be added back in later. Currently having managment form data has been tampered with in the tests.